% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_within.R
\name{are_within}
\alias{are_within}
\title{Are points within a set distance?}
\usage{
are_within(lat, lon, radius = "1 km", lambda0 = NULL)
}
\arguments{
\item{lat, lon}{Coordinates of points.}

\item{radius}{Radius in kiometres.}

\item{lambda0}{Passed to the sinusoidal projection. Defaults to the
median of \code{lon}. Use to indicate the longitude at which distortion
should be zero (and smaller the the closer the longitude).}
}
\value{
Logical vector \code{x} such that,
for every \code{(lat[i],lon[i])}, \code{x[i]} is
\code{TRUE} iff there is some \code{j} such that
\code{dist_geo(lat[i], lon[i], lat[j], lon[j]) < r}.
}
\description{
Are points within a set distance?
}
