% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cj2.R
\name{cj2}
\alias{cj2}
\title{Two dimensional cross join}
\usage{
cj2(
  u,
  v,
  as_data_table = TRUE,
  new_cols = c("id1", "id2", "u1", "v1", "u2", "v2"),
  nThread = 1L
)
}
\arguments{
\item{u, v}{Two vectors.}

\item{as_data_table}{(bool) Return a \code{data.table} (as opposed to a list, the default)?}

\item{new_cols}{The name of the list/\code{data.table}.}

\item{nThread}{The number of thread to use.}
}
\value{
A list (or \code{data.table} if requested), representing the
expanded grid as if the data cross joined.
}
\description{
Two dimensional cross join
}
