% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engrid.R
\name{engrid}
\alias{engrid}
\title{Engrid}
\usage{
engrid(x, y, r)
}
\arguments{
\item{x, y}{\code{double(N)} Coordinates.}

\item{r}{Radius of each cell in the grid.}
}
\description{
Create a grid of radius \code{r} over the range of \code{x, y},
index them and return the gridpoint of each point.
}
