% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{group_contiguous}
\alias{group_contiguous}
\title{Gaps}
\usage{
group_contiguous(x, r)
}
\arguments{
\item{x}{An integer vector.}

\item{r}{The minimum radius such that \eqn{|x_i - x_j| > r} implies
\eqn{x_i,x_j} are not in the same group. Will be coerced to as.integer(ceil(r)).}
}
\value{
For each element \code{x} a group index.
}
\description{
For an integer vector (such as time) find gaps, and
return a index based "sufficiently continguous" groups.
}
