% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_latlons.R
\name{plausible_latloncols}
\alias{plausible_latloncols}
\title{Which names are plausibly latlons}
\usage{
plausible_latloncols(DT, vname_DT = "DT")
}
\arguments{
\item{DT}{An object with names.}

\item{vname_DT}{The name of the table to return in error messages.}
}
\value{
A character vector of length-2, the names of \code{DT} that
will be plausibly be lat,lon.
}
\description{
Which names are plausibly latlons
}
