% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2time.R
\name{string2duration}
\alias{string2duration}
\title{Convert a string into a duration in seconds}
\usage{
string2duration(x)
}
\arguments{
\item{x}{A string, like "2 minutes" representing a duration}
}
\value{
An integer, that duration in seconds.
}
\description{
Convert a string into a duration in seconds
}
