% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units2km.R
\name{units2km}
\alias{units2km}
\title{Convert a distance as a string to a number}
\usage{
units2km(input)
}
\arguments{
\item{input}{The input to standardize.}
}
\value{
If \code{input} is numeric, then it is returned, immediately.

Otherwise, if a string, the distance in kilometres
the string appears to represent.
}
\description{
Standardizes input like \code{"500 m"}, \code{"0.5 km"} and
\code{0.5} (numeric) to a single double, (in this case, \code{0.5}). Useful
so the user can clearly express the units, rather than entering a bare
numeric.
}
\examples{
units2km(5) 
units2km("5 km")
units2km("50 m")

}
