% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_within_dist_dura.R
\name{which_within_dist_dura}
\alias{which_within_dist_dura}
\title{Find matches by distance and time}
\usage{
which_within_dist_dura(
  x,
  Data,
  x_column,
  time_column,
  distance = "100m",
  duration = 3600,
  option = 0L,
  Ion = getOption("whichWithin.Ion", 10)
)
}
\arguments{
\item{x}{An integer vector, an index of \code{Data}, to restrict the values
of \code{orig} in the result.}

\item{Data}{A \code{data.table} containing columns for latitude and longitude
(see \code{\link{plausible_latloncols}}), as well as columns \code{time_column}
and \code{x_column}.}

\item{x_column, time_column}{Names/indices of \code{Data} corresponding to integer
vectors for filtering and determining time-matching.}

\item{distance}{A string, passed to \code{\link{units2km}}: the maximum distance between matches.}

\item{duration}{The maximum difference in time in seconds.}

\item{option}{An integer. Only \code{option = 0L} is allowed currently.
\describe{
\item{0}{Restricts values of \code{orig}; any exact matches mean approximate matches are ignored.}
\item{1}{Same as \code{0} but \code{dest} is restricted too.}
\item{2}{Exact matches do not preclude approximate matches}
}}

\item{Ion}{Used to set the expected number of rows required in the output. Memory is allocated
in large chunks within this function. If you find yourself running out of memory
(with an error message referring to a \code{realloc}) you can tweak this parameter.
A slight overestimate is much better than a slight underestimate as an insufficient
initial allocation will cause a large reallocation when required. (e.g. if
100M rows are required and \code{Ion = 99e6} then approximately 160M rows will
be reallocated).}
}
\value{
A list of two integer vectors, named \code{orig} and \code{dest},
the indices of \code{Data} corresponding to pairs of observations that
are within \code{distance} and \code{duration} of each other. Will be a \code{data.table}
if \code{length(orig) <= INT_MAX}.
}
\description{
Find matches by distance and time
}
