% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_within_x1d.R
\name{which_within_x1d}
\alias{which_within_x1d}
\title{Which exact and within}
\usage{
which_within_x1d(x, y, r, option = 1L)
}
\arguments{
\item{x}{A vector to identify exact matches.}

\item{y}{A vector to identify near matches.}

\item{r}{The radius within which matches of \code{y} occur.}

\item{option}{For backwards compatible changes.}
}
\value{
Considering \code{x, y} as coordinates, we define the
binary relation
\deqn{(x[i], y[i]) \mathrm{within} (x[j], y[j])}
if \eqn{x[i] = y[j]} and \eqn{|y[i] - y[j]| < r}).

A list of two elements, \code{orig} and \code{dest} giving the
indices of \code{x, y} that satisfy the binary relation above.
}
\description{
Used to identify pairs that have exact matches in one
vector and near matches in another.
}
