% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichWithin-package.R
\docType{package}
\name{whichWithin-package}
\alias{whichWithin}
\alias{whichWithin-package}
\title{whichWithin}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
